* using uint64_t for mem_sizes as default

+++ src/iofwd/WriteRequest.hh
@@ -16,7 +16,7 @@ public:
      zoidfs::zoidfs_handle_t * handle;
      uint32_t mem_count;
      char ** mem_starts;
-     uint32_t * mem_sizes;
+     uint64_t * mem_sizes;
      uint32_t file_count;
      uint64_t * file_starts;
      uint64_t * file_sizes;
